requirements.txt
controlnet-aux==0.0.3
diffusers==0.16.1
fastapi==0.111.0
gradio==4.31.0
huggingface-hub==0.23.0
imgviz==1.7.5
ipython==7.34.0
labelme==5.4.1
loguru==0.7.2
matplotlib==3.7.1
numpy==1.25.2
omegaconf==2.3.0
openai==0.28.0
opencv-python>=4.8.0.76
peft==0.7.1
piexif==1.1.3
Pillow==9.4.0
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-socketio==5.11.2
rich==13.7.1
safetensors==0.4.3
tensorflow==2.15.0
torch==2.2.2
transformers==4.27.4
typer==0.12.3
uvicorn==0.29.0
yacs==0.1.8
dramatiq==1.17.1
flask
sqlalchemy